Cherry pick from vision

+++ src/main/webapp/WEB-INF/jsp/finalexamreport.jsp
@@ -1,18 +1,19 @@
 <%--
-    Document   : studentsdetailsbonafide
-    Created on : Jul 09, 2016, 12:10:28 AM
+    Document   : progress report
+    Created on : SEP 23, 2018, 5:52:28 PM
     Author     : Musaib
 --%>
 
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+<%@ taglib prefix="fn" uri="http://java.sun.com/jsp/jstl/functions"  %>
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
     "http://www.w3.org/TR/html4/loose.dtd">
 
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Article 371</title>
+<title>Progress Card</title>
 <link rel="stylesheet" href="/school/css/datePicker/jquery-ui-1.8.18.custom.css">
 <link rel="stylesheet" href="/school/css/datePicker/demos.css">
 <style type="text/css">
@@ -358,9 +359,11 @@
 
 <script type="text/javascript" src="/school/js/datetimepicker_css.js"></script>
 <script type="text/javascript">
-	function searchStudentsForBonafide() {
+	function searchForReports() {
+		
+		
 		var form1 = document.getElementById("form1");
-		form1.action = "/school/DocumentsProcess/searchStudentsForArticle";
+		form1.action = "/school/MarksDetailsProcess/searchForFinalReport";
 		form1.method = "POST";
 		form1.submit();
 
@@ -369,7 +372,7 @@
 	$(function() {
 
 		$("#search").button().click(function() {
-			searchStudentsForBonafide();
+			searchForReports();
 		});
 		
 
@@ -379,23 +382,26 @@
 
 		$("#tabs").tabs();
 
-		$("#save").button().click(function() {
-			addDepartment();
-		});
-		/* $("#effect").hide(); */
-
 	});
 	
 	$(function() {
-		$("#export").button({
+		$("#generatefinalexamreport").button({
 			icons : {
 				primary : "ui-icon-trash"
 			}
-		}).click(function() {
-			deleteRecords();
-			return false;
-
 		});
+		$("#delete").button({
+            icons:{
+                primary: "ui-icon-trash"
+            }
+        }).click(function(){
+        	if(confirm('Are you sure,you want to delete?')){
+        		deleteRecords();	
+        	}
+            
+            return false;
+
+        });
 		
 		$('#chckHead').click(function() {
 			var length = $('.chcktbl:checked').length;
@@ -426,181 +432,11 @@
 			}
 		});
 
-		$("#go").button()
 
 	});
 	
-	function deleteFeesStamp(){
-		var form1 = document.getElementById("form1");
-		form1.action = "/school/StampFeesProcess/delete";
-		form1.method = "POST";
-		form1.submit();
-	}
+	
 </script>
-
-
-<script type="text/javascript">
-        
-        function calculateGrandTotal() {
-            var sum = 0.0;
-            var column2 = $('.feesFullAmount')
-            jQuery.each(column2,function(){
-                sum += parseFloat($(this).val());
-            });
-            
-            $('#feesTotalAmount').val(sum.toPrecision(6));
-
-        }
-        $(document).ready(function() {
-            
-            
-            $("#dataTable").keyup(function(){
-                
-                var sum = 0.0;
-                var totalSum=0.0;
-                var column2 = $('.feesFullAmount')
-                jQuery.each(column2,function(){
-                    sum += parseFloat($(this).val());
-                });
-                
-                $('#feesTotalAmount').val(sum.toPrecision(6));
-                
-            });
-            $("#dataTable").click(function(){
-                
-                var sum = 0.0;
-                var totalSum=0.0;
-                var column2 = $('.feesFullAmount')
-                jQuery.each(column2,function(){
-                    sum += parseFloat($(this).val());
-                });
-                
-                $('#feesTotalAmount').val(sum.toPrecision(6));
-               
-            });
-
-
-        });
-        
-    
-        
-        $(function() {
-            
-            var addFeesCatButtonID="#addFeesCat";
-            var removeFeesCatButtonID="#removeFeesCat";
-            $( addFeesCatButtonID )
-            .button({
-                icons: {
-                    primary: "ui-icon-plus"
-                }
-            })
-            .click(function() {
-                addRow();
-                return false;
-            });
-            $(removeFeesCatButtonID)
-            .button({
-                icons: {
-                    primary: "ui-icon-minus"
-                }
-            })
-            .click(function() {
-                deleteRow('dataTable');
-                return false;
-            });            
-
-        });
-        
-        function SelectAll(id)
-        {
-        	
-            document.getElementById("feesCount_"+id).focus();
-            document.getElementById("feesCount_"+id).select();
-        }
-
-        function calculate(value2) {
-
-        	var feesCount=document.getElementById("feesCount_"+value2).value;
-        	
-        	if(feesCount === ''){
-        		
-        		document.getElementById("feesCount_"+value2).value = 1;
-        	}
-        	
-            //var val1=value1.value;
-            var feesCat=document.getElementById("hiddenfees_amount_"+value2).value;
-            var feesCount=document.getElementById("feesCount_"+value2).value;
-            var final1=document.getElementById("hiddenfees_full_amount_"+value2);
-            
-                final1.value=(feesCat*feesCount).toPrecision(6);
-           
-        }
-       
-    function selectAllRow(tableID){
-        var table = document.getElementById(tableID);
-        var rowCount = table.rows.length;
-        if(rowCount==1){
-            var row = table.rows[0];
-            var chkbox = row.cells[0].childNodes[0];
-            chkbox.checked=false;
-            alert('No records to select');
-        }
-        for(var i=1; i<rowCount; i++) {
-            var row = table.rows[i];
-            var chkbox = row.cells[0].childNodes[0];
-            chkbox.checked=true;
-        }
-    }
-    
-    function deleteRow(tableID) {
-        try {
-            var table = document.getElementById(tableID);
-            var rowCount = table.rows.length;
-            if(rowCount==1){
-                alert('No records to delete');
-            }
-            for(var i=1; i<rowCount-1; i++) {
-                var row = table.rows[i];
-                var chkbox = row.cells[0].childNodes[0];
-                if(null != chkbox && true == chkbox.checked) {
-                    table.deleteRow(i);
-                    rowCount--;
-                    i--;
-                }
-            }
-           
-            
-            var sum = 0.0;
-            var totalSum=0.0;
-            var column2 = $('.feesAmount')
-            jQuery.each(column2,function(){
-                sum += parseFloat($(this).val());
-            });
-            totalSum=sum;
-            
-            $('#feesTotalAmount').val(totalSum.toPrecision(6));
-            
-            calculateGrandTotal();
-            //$('#grandTotalAmount').val(0);
-        }catch(e) {
-            alert(e);
-        }
-    }
-    
-	function fileCheck() {
-		
-		if (document.getElementById("fileName").value.length == 0)
-
-		{
-			document.getElementById("fileName").style.background = 'red';
-			alert("Enter The File Name ");
-		}
-
-	}
-
-        </script>
-
-
 
 </head>
   <%
@@ -620,7 +456,7 @@ for(Cookie cookie : cookies){
 }
 %>
 <body>
-	<form id="form1" action="/school/DocumentsProcess/GenerateArticle" method="POST">
+	<form id="form1" action="/school/MarksDetailsProcess/generatefinalexamReport" method="POST">
 		<!-- <div style="height: 28px">
 			<button id="add">Add Department</button>
 			<br />
@@ -629,7 +465,7 @@ for(Cookie cookie : cookies){
 		<div id="effect" class="ui-widget-content ui-corner-all">
 			<div id="tabs">
 				<ul>
-					<li><a href="#tabs-1">Article 371 Certificate</a></li>
+					<li><a href="#tabs-1">Final Exam Report</a></li>
 
 				</ul>
 				<div id="tabs-1">
@@ -640,7 +476,7 @@ for(Cookie cookie : cookies){
 							<td class="alignRightFields">Name &nbsp;</td>
 							<td width="12%" align="left"><label> <input
 									name="namesearch" type="text" class="myclass" id="namesearch"
-									size="36"">
+									size="40">
 							</label></td>
 							
 						</tr>
@@ -654,38 +490,41 @@ for(Cookie cookie : cookies){
 						<tr>
 							<td class="alignRightFields">Class &nbsp;</td>
 							<td width="70%"><label> <select name="classsearch"
-									id="classsearch" style="width: 150px">
+									id="classsearch" style="width: 128px;">
 										<option selected></option>
 										<c:forEach items="${classdetailslist}" var="classdetailslist">
-										<c:if test="${(classdetailslist.classdetails != '')}">
-											<option value="${classdetailslist.classdetails}" >
-												<c:out value="${classdetailslist.classdetails}" />
-											</option>
-										</c:if>	
+											<c:if test="${(classdetailslist.classdetails != '')}">
+												<option value="${classdetailslist.classdetails}">
+													<c:out value="${classdetailslist.classdetails}" />
+												</option>
+											</c:if>
 										</c:forEach>
 								</select>
 
 							</label> <label> <select name="secsearch" id="secsearch"
-									style="width: 120px">
+									style="width: 128px;">
 										<option selected></option>
 
 										<c:forEach items="${classdetailslist}" var="classdetailslist">
-										<c:if test="${(classdetailslist.section != '')}">
-											<option value="${classdetailslist.section}">
-												<c:out value="${classdetailslist.section}" />
-											</option>
-										</c:if>	
+											<c:if test="${(classdetailslist.section != '')}">
+												<option value="${classdetailslist.section}">
+													<c:out value="${classdetailslist.section}" />
+												</option>
+											</c:if>
 										</c:forEach>
-
 								</select>
-							</label>
+							</label></td>
 						</tr>
 
-						<tr>
+
+				<tr>
 							<td><br /></td>
-
 						</tr>
 
+
+							<td><br /></td>
+						</tr>
+						
 						<tr>
 
 							<td width="30%" class="alignRight"></td>
@@ -695,12 +534,35 @@ for(Cookie cookie : cookies){
 								<button id="search">Search</button>
 							</td>
 						</tr>
-
-
+						
+						
 						<tr>
 							<td><br /></td>
-						</tr>
 
+						</tr>
+						
+						
+
+						
+
+
+						
+						<tr>
+						<td> Exam Name </td>
+							<td><select name="examname"
+									id="examname" style="width: 128px;">
+										<option selected></option>
+										<c:forEach items="${listExam}" var="listExam">
+											<c:if test="${(listExam.examname != '')}">
+												<option value="${listExam.examname}">
+													<c:out value="${listExam.examname}" />
+												</option>
+											</c:if>
+										</c:forEach>
+								</select></td>
+						</tr>
+						
+						
 					</table>
 					
 					
@@ -721,13 +583,11 @@ for(Cookie cookie : cookies){
 				<thead>
 					<tr>
 						<th class="headerText"><input type="checkbox" id="chckHead" /></th>
+						<th title="click to sort" class="headerText">UID</th>
 						<th title="click to sort" class="headerText">Admission Number</th>
 						<th title="click to sort" class="headerText">Name&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</th>
-						<th title="click to sort" class="headerText">Class &
-							Sec&nbsp;</th>
-						<th title="click to sort" class="headerText">Father's Name</th>
-						<th title="click to sort" class="headerText">Mother's Name</th>	
-						<th title="click to sort" class="headerText">Admission Date</th>
+						<th title="click to sort" class="headerText">Class</th>
+						
 
 
 
@@ -735,7 +595,7 @@ for(Cookie cookie : cookies){
 				</thead>
 
 				<tbody>
-					<c:forEach items="${searchStudentList}" var="Parents">
+					<c:forEach items="${searchStudentList}" var="Parents" >
 
 						<tr class="trClass" style="border-color: #000000" border="1"
 							cellpadding="1" cellspacing="1">
@@ -743,27 +603,32 @@ for(Cookie cookie : cookies){
 								id="<c:out value="${Parents.student.sid}"/>" class="chcktbl"
 								name="studentIDs"
 								value="<c:out value="${Parents.student.sid}"/>" /></td>
-							<td class="dataText"><c:out
+								<td class="dataTextInActive"><a class="dataTextInActive"
+								><c:out
+										value="${Parents.student.studentexternalid}" /></a></td>
+								<td class="dataTextInActive"><a class="dataTextInActive"
+								><c:out
 										value="${Parents.student.admissionnumber}" /></a></td>
 							<td class="dataText"><c:out value="${Parents.student.name}" /></td>
-							<td class="dataText"><c:out
-									value="${Parents.student.classstudying}" /></td>
-							<td class="dataText"><c:out
-									value="${Parents.fathersname}" /></td>
-							<td class="dataText"><c:out
-									value="${Parents.mothersname}" /></td>
-							<td class="dataText"><c:out
-									value="${Parents.student.admissiondate}" /></td>
+							<td class="dataText">
+							<input type="hidden" id="examclass" name="examclass"
+								value="<c:out value="${Parents.student.classstudying}"/>" />
+							 <c:forEach var="splt" items="${fn:split(Parents.student.classstudying,'--')}">
+						    ${splt} 
+							</c:forEach>
+							</td>
+							
 						</tr>
 					</c:forEach>
 				</tbody>
 				<tfoot>
 					<tr>
-													
-						<td class="footerTD" colspan="2"> &nbsp;
-						<input value="Generate Certificate"
-							type="submit" id="export"/></td>
+					
+						<td class="footerTD" colspan="2"><input value="Generate Report"
+							type="submit" id="generatefinalexamreport" />
 							
+							<!-- <input value="Delete Stamp Fees"
+							type="submit" id="deleteStamp" /> --></td>
 							
 
 					</tr>
